+++ packages/mint-components/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.9.4] - 2024-07-04
+
+### Changed
+
+- `@saasquatch/component-boilerplate` package bump
+- \<sqm-leaderboard>
+  - No longer batches leaderboard query to avoid the batch from failing when permissions are not set
+
 ## [1.9.3] - 2024-07-03
 
 ### Changed
@@ -15,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added required feature flag
 - \<sqm-portal-footer>
   - Added required feature flag
-  
+
 ## [1.9.2] - 2024-06-28
 
 ### Fixed
@@ -989,7 +997,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - \<sqm-popup-container>
   - \<sqm-stencilbook>
 
-[unreleased]: https://github.com/saasquatch/program-tools/compare/mint-components@1.9.2...HEAD
+[unreleased]: https://github.com/saasquatch/program-tools/compare/mint-components@1.9.4...HEAD
+[1.9.4]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.9.4
+[1.9.3]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.9.3
 [1.9.2]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.9.2
 [1.9.1]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.9.1
 [1.9.0]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.9.0